Move tracking logic to utils
